# docs/dev/api/openapi.yaml
openapi: 3.1.0
info:
  title: Auth API
  description: 'Authentication API: accounts, sessions, second factors and client applications. Issues ES256 access tokens and rotating refresh tokens, and publishes the JWKS resource servers verify them with.'
  license:
    name: MIT
  version: 2.0.1
servers:
- url: http://localhost:3000
  description: Local development
paths:
...
  /users/me:
    get:
      tags:
      - account
      operationId: me
      responses:
        '200':
          description: The caller's profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: '`first_party_session_required`: a token delegated to a client or issued for a personal access token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
    delete:
      tags:
      - account
      operationId: delete_account
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: 'null'
              - $ref: '#/components/schemas/DeleteAccountRequest'
      responses:
        '204':
          description: Account deleted
        '400':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Recent re-authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '413':
          description: Body larger than 64 KB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '415':
          description: Body is not JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
...
  /users/me/email/confirm:
    post:
      tags:
      - email-change
      operationId: confirm_new_email
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmNewEmailRequest'
        required: true
      responses:
        '204':
          description: Address changed; other sessions revoked
        '400':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: '`first_party_session_required`: a token delegated to a client or issued for a personal access token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Address taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '413':
          description: Body larger than 64 KB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '415':
          description: Body is not JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
  /users/me/email/start:
    post:
      tags:
      - email-change
      operationId: start_email_change
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: 'null'
              - $ref: '#/components/schemas/CurrentPasswordRequest'
      responses:
        '200':
          description: Code sent to the current address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowTokenResponse'
        '400':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Recent re-authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '413':
          description: Body larger than 64 KB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '415':
          description: Body is not JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
  /users/me/email/submit:
    post:
      tags:
      - email-change
      operationId: submit_new_email
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitNewEmailRequest'
        required: true
      responses:
        '204':
          description: Code sent to the new address
        '400':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: '`first_party_session_required`: a token delegated to a client or issued for a personal access token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: Address taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '413':
          description: Body larger than 64 KB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '415':
          description: Body is not JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
  /users/me/email/verify-current:
    post:
      tags:
      - email-change
      operationId: verify_current_email
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyCurrentEmailRequest'
        required: true
      responses:
        '204':
          description: Current address confirmed
        '400':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: '`first_party_session_required`: a token delegated to a client or issued for a personal access token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '413':
          description: Body larger than 64 KB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '415':
          description: Body is not JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
  /users/me/export:
    get:
      tags:
      - account
      operationId: export_data
      responses:
        '200':
          description: 'Everything stored about the account, as a JSON download: profile, roles, sessions, second factors (no secrets), devices, sign-in attempts and security history'
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: Recent re-authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
...
  /users/me/locale:
    patch:
      tags:
      - account
      operationId: change_locale
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeLocaleRequest'
        required: true
      responses:
        '204':
          description: Locale changed
        '400':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '401':
          description: Missing, invalid or revoked access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '403':
          description: '`first_party_session_required`: a token delegated to a client or issued for a personal access token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '413':
          description: Body larger than 64 KB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '415':
          description: Body is not JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '422':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '429':
          description: Rate limited; see Retry-After
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: A dependency is unavailable or the request timed out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
      security:
      - bearer: []
...
components:
  schemas:
...
    ChangeLocaleRequest:
      type: object
      required:
      - locale
      properties:
        locale:
          type: string
...
    ConfirmNewEmailRequest:
      type: object
      required:
      - flow_token
      - code
      properties:
        code:
          type: string
        flow_token:
          type: string
...
    CurrentPasswordRequest:
      type: object
      description: |-
        Optional body for sensitive actions that accept the current password in
        place of a recent re-authentication.
      properties:
        current_password:
          type:
          - string
          - 'null'
    DeleteAccountRequest:
      type: object
      properties:
        current_password:
          type:
          - string
          - 'null'
...
    ErrorBody:
      type: object
      description: Body of every error response.
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: 'Stable and machine-readable: `invalid_credentials`, `reauthentication_required`, ...'
        message:
          type: string
          description: Human-readable explanation; may change between versions.
...
    FlowTokenResponse:
      type: object
      required:
      - flow_token
      properties:
        flow_token:
          type: string
...
    SubmitNewEmailRequest:
      type: object
      required:
      - flow_token
      - new_email
      properties:
        flow_token:
          type: string
        new_email:
          type: string
...
    UserResponse:
      type: object
      required:
      - id
      - username
      - email
      - status
      - preferred_locale
      - created_at
      properties:
        created_at:
          type: integer
          format: int64
        email:
          type: string
        email_verified_at:
          type:
          - integer
          - 'null'
          format: int64
        id:
          type: string
          format: uuid
        last_login_at:
          type:
          - integer
          - 'null'
          format: int64
        preferred_locale:
          type: string
        status:
          type: string
        username:
          type: string
    VerifyCurrentEmailRequest:
      type: object
      required:
      - flow_token
      - code
      properties:
        code:
          type: string
        flow_token:
          type: string
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...